# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bipper api",
        "title": "Bipper Api",
        "contact": {
            "name": "Batleforc",
            "url": "https://weebo.fr",
            "email": "maxleriche.60@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login body",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route.LoginBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login return",
                        "schema": {
                            "$ref": "#/definitions/route.LoginReturn"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Logout user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "parameters": [
                    {
                        "description": "Logout body",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route.LogoutBody"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register User",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register User",
                "parameters": [
                    {
                        "description": "Register body",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route.RegisterBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Register return",
                        "schema": {
                            "$ref": "#/definitions/route.RegisterReturn"
                        }
                    }
                }
            }
        },
        "/auth/renew": {
            "post": {
                "description": "Renew Token via refresh token",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Renew Token",
                "parameters": [
                    {
                        "description": "Renew body",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/route.RenewTokenBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Renew return",
                        "schema": {
                            "$ref": "#/definitions/route.RenewTokenReturn"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "description": "Get user",
                "tags": [
                    "Auth"
                ],
                "summary": "Get user",
                "responses": {
                    "200": {
                        "description": "user return",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "model.Channel": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Message"
                    }
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "integer"
                },
                "passKey": {
                    "description": "crypted field (can be regenerated) and empty if public",
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "private": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ChannelUser"
                    }
                }
            }
        },
        "model.ChannelUser": {
            "type": "object",
            "properties": {
                "canMod": {
                    "type": "boolean"
                },
                "canRead": {
                    "type": "boolean"
                },
                "canSend": {
                    "type": "boolean"
                },
                "channelID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "model.Message": {
            "type": "object",
            "properties": {
                "channelID": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "model.Token": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "channels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ChannelUser"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "myChannels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Channel"
                    }
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "pseudo": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "tokens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Token"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "route.LoginBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "route.LoginReturn": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "pseudo": {
                    "type": "string"
                },
                "renew_token": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "route.LogoutBody": {
            "type": "object",
            "properties": {
                "renew_token": {
                    "type": "string"
                }
            }
        },
        "route.RegisterBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "pseudo": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "route.RegisterReturn": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "registered": {
                    "type": "boolean"
                }
            }
        },
        "route.RenewTokenBody": {
            "type": "object",
            "properties": {
                "renew_token": {
                    "type": "string"
                }
            }
        },
        "route.RenewTokenReturn": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "pseudo": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        }
    }
}